refactor: move src/util/historyFile->src/history/historyFile
